7th Jan/requirements.txt
fastapi
uvicorn
requests
qdrant-client
python-dotenv
